fix: transform light direction into camera space
The light direction vector was not being transformed into camera space,
which caused incorrect lighting calculations when the camera rotated.
Now properly transforming the light direction using transform_vertex_normal
with camera rotation to ensure consistent lighting in camera space.

+++ src/kernel/transforming_mod.py
@@ -130,7 +130,7 @@ def transform_vertex(vertex: np.ndarray,
     
     # Return the transformed vertex as a new array
     # This ensures the original vertex array is not modified
-    return np.array([x, y, z])
+    return np.array([x, y, z], dtype=np.float32)
 
 
 @njit(cache=True)
@@ -221,7 +221,7 @@ def transform_vertex_normal(normal: np.ndarray,
         y /= length
         z /= length
     
-    return np.array([x, y, z])
+    return np.array([x, y, z], dtype=np.float32)
 
 
 @njit(cache=True)
